requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
youtube-transcript-api>=0.6.0
google-genai>=1.0.0
python-dotenv>=1.0.0
